rpm: fix processing of metafiles
The patch fixes several problems while metafiles processing:
* removing the created metafile if the names of old and new metafiles ("filelists" and "primary") are the same
* crashing when one (or more) files specified in "repomd" are absent

Fixes #38

+++ rpmrepo.py
@@ -799,12 +799,22 @@ def parse_metafiles(storage):
         filelists, primary, revision = parse_repomd(data)
 
         initial_filelists = filelists['location']
-        data = storage.read_file(initial_filelists)
-        filelists = parse_filelists(gunzip_bytes(data))
+        # The file can be specified in repomd.xml but doesn't exist.
+        if storage.exists(initial_filelists):
+            data = storage.read_file(initial_filelists)
+            filelists = parse_filelists(gunzip_bytes(data))
+        else:
+            initial_filelists = None
+            filelists = {}
 
         initial_primary = primary['location']
-        data = storage.read_file(initial_primary)
-        primary = parse_primary(gunzip_bytes(data))
+        # The file can be specified in repomd.xml but doesn't exist.
+        if storage.exists(initial_primary):
+            data = storage.read_file(initial_primary)
+            primary = parse_primary(gunzip_bytes(data))
+        else:
+            initial_primary = None
+            primary = {}
 
     return filelists, primary, revision, initial_filelists, initial_primary
 
@@ -890,9 +900,18 @@ def update_repo(storage, sign, tempdir, force=False):
     storage.write_file(primary_name, primary_gz)
     storage.write_file('repodata/repomd.xml', repomd_str.encode('utf-8'))
 
-    if initial_filelists:
+    # Here we are deleting few old metafiles.
+    # The difference in names between the old and new files (in case the new
+    # packages were not added to the repository) exists because part of the
+    # name is the sha256 hash from  the".gz" files, wich includes information
+    # about time when it was created (in seconds (float)). In my opinion, this
+    # difference is not obvious on the one hand and may cease to exist on the
+    # other hand (if we start to set timestamp according to a different logic
+    # or if the hashsum is removed from the filename).
+    # Let's check the names for equivalence.
+    if initial_filelists and initial_filelists != filelists_name:
         storage.delete_file(initial_filelists)
-    if initial_primary:
+    if initial_primary and initial_primary != primary_name:
         storage.delete_file(initial_primary)
 
     if sign:
